task_1 completed

+++ src/DataTypes.java
@@ -3,8 +3,8 @@ import java.util.List;
 public class DataTypes {
     // TODO TASK 1: fix this code so that it passes the test in DataTypesTest.java
     public static long sum(List<Integer> numbers) {
-
-        int s = 0;
+        //int s = 0 doesn't work
+        long s = 0;
         // below is a "foreach" loop which iterates through numbers
         for (int x : numbers) {
             s += x;
